Common: Move generic vertex buffer code from d3d12 backend
